Organize project based on pages
